requirements.txt
Flask>=2.0
requests>=2.25
<<<<<<< HEAD
SQLAlchemy>=1.4
mysql-connector-python>=8.0
python-dotenv>=0.19
=======

